requirements.txt
asgiref==3.8.1
contourpy==1.3.0
cycler==0.12.1
Django==5.1.2
djangorestframework==3.15.2
drf-yasg==1.21.8
fonttools==4.54.1
inflection==0.5.1
kiwisolver==1.4.7
matplotlib==3.9.2
numpy==2.1.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
seaborn==0.13.2
six==1.16.0
sqlparse==0.5.1
squarify==0.4.4
tzdata==2024.2
uritemplate==4.1.1
wordcloud==1.9.3
